+++ features/ShowModVersions/ui/index.ts
@@ -1,0 +1,3 @@
+import ShowModVersions from './ShowModVersions.vue'
+
+export { ShowModVersions }
