+++ pages/HomePage/model/index.ts
@@ -1,1 +1,3 @@
-export { }
+import useFilters from './useFilters'
+
+export { useFilters }
